Fix conversion during lookup so we don't error out on sets
When using the "overlap" lookup, code would crash with a "can't adapt
type: set" when provided a set instead of a list.  It would also crash
for an exact match lookup.

Somehow, the contains and contained_by lookups already work (and of
course those were the only ones I tested...).  Maybe they convert back
and forth between sets already?  Couldn't find the code for this,
though.